Move some components into shared file
