+++ app/src/main/java/data/Common.kt
@@ -1,0 +1,17 @@
+package data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Artifact(
+    var name: Int,
+    var rarity: Int,
+    var level: Int,
+    var stones: List<Stone>
+)
+
+@Serializable
+data class Stone(
+    var name: Int,
+    var level: Int
+)
