+++ app/src/main/java/data/VirtueInfo.kt
@@ -7,8 +7,9 @@ data class VirtueInfo(
     var stateId: String = "",
     var resets: Int = 0,
     var shifts: Int = 0,
-    var soulEggs: String = "",
+    var nextShiftCost: String = "",
     var totalTruthEggs: String = "",
+    var totalPendingTruthEggs: String = "",
     var eggId: Int = 0,
     var population: String = "",
     var lastBackupDate: Double = 0.0,
@@ -27,20 +28,6 @@ data class VirtueFarmInfo(
     val truthEggs: Int = 0,
     val pendingTruthEggs: Int = 0,
     val eggsDelivered: Double = 0.0
-)
-
-@Serializable
-data class Artifact(
-    var name: Int,
-    var rarity: Int,
-    var level: Int,
-    var stones: List<Stone>
-)
-
-@Serializable
-data class Stone(
-    var name: Int,
-    var level: Int
 )
 
 @Serializable
